Move: NotificationType Entity 패키지로 이동
