+++ src/main/java/com/sesac/joinflex/domain/notification/dto/response/NotificationEventResponse.java
@@ -1,6 +1,6 @@
 package com.sesac.joinflex.domain.notification.dto.response;
 
-import com.sesac.joinflex.domain.notification.type.NotificationType;
+import com.sesac.joinflex.domain.notification.entity.NotificationType;
 import lombok.Builder;
 import lombok.Getter;
