+++ src/main/java/com/sesac/joinflex/domain/notification/entity/NotificationType.java
@@ -1,0 +1,8 @@
+package com.sesac.joinflex.domain.notification.entity;
+
+public enum NotificationType {
+    FRIEND_REQUEST,
+    FRIEND_ACCEPT,
+    FRIEND_REJECT,
+
+}
