--- src/main/java/com/sesac/joinflex/domain/notification/type/NotificationType.java
@@ -1,7 +1,0 @@
-package com.sesac.joinflex.domain.notification.type;
-
-public enum NotificationType {
-    FRIEND_REQUEST,
-    FRIEND_ACCEPT,
-    FRIEND_REJECT
-}
